+++ packages/compiler/test/semantic-walker.test.ts
@@ -17,15 +17,23 @@ import {
 import {
   getProperty,
   navigateProgram,
+  navigateType,
   navigateTypesInNamespace,
 } from "../src/core/semantic-walker.js";
-import { TestHost, createTestHost } from "../src/testing/index.js";
+import {
+  BasicTestRunner,
+  TestHost,
+  createTestHost,
+  createTestRunner,
+} from "../src/testing/index.js";
 
 describe("compiler: semantic walker", () => {
   let host: TestHost;
+  let runner: BasicTestRunner;
 
   beforeEach(async () => {
     host = await createTestHost();
+    runner = await createTestRunner();
   });
 
   function createCollector(customListener?: SemanticNodeListener) {
@@ -137,6 +145,100 @@ describe("compiler: semantic walker", () => {
 
     return result;
   }
+
+  it("finds derived models", async () => {
+    const { Bird } = (await runner.compile(`
+      namespace Test;
+
+      @discriminator("kind")
+      @test 
+      model Bird {
+        kind: string;
+        wingspan: int32;
+      }
+
+      model SeaGull extends Bird {
+        kind: "seagull";
+      }
+
+      model Sparrow extends Bird {
+        kind: "sparrow";
+      }
+
+      model Goose extends Bird {
+        kind: "goose";
+      }
+
+      model Eagle extends Bird {
+        kind: "eagle";
+        friends?: Bird[];
+        hate?: Record<Bird>;
+        partner?: Bird;
+      }
+      `)) as { Bird: Model };
+
+    const visitedModels: Model[] = [];
+    navigateType(
+      Bird,
+      {
+        model(model) {
+          visitedModels.push(model);
+        },
+      },
+      { includeTemplateDeclaration: false, visitDerivedTypes: true },
+    );
+
+    const expectedModels = ["Bird", "SeaGull", "Sparrow", "Goose", "Eagle"];
+    strictEqual(
+      expectedModels.every((element) => visitedModels.map((m) => m.name).includes(element)),
+      true,
+    );
+  });
+
+  it("doesn't visit derived models without the option", async () => {
+    const { Bird } = (await runner.compile(`
+      namespace Test;
+
+      @discriminator("kind")
+      @test 
+      model Bird {
+        kind: string;
+        wingspan: int32;
+      }
+
+      model SeaGull extends Bird {
+        kind: "seagull";
+      }
+
+      model Sparrow extends Bird {
+        kind: "sparrow";
+      }
+
+      model Goose extends Bird {
+        kind: "goose";
+      }
+
+      model Eagle extends Bird {
+        kind: "eagle";
+        friends?: Bird[];
+        hate?: Record<Bird>;
+        partner?: Bird;
+      }
+      `)) as { Bird: Model };
+
+    const visitedModels: Model[] = [];
+    navigateType(
+      Bird,
+      {
+        model(model) {
+          visitedModels.push(model);
+        },
+      },
+      { visitDerivedTypes: false },
+    );
+
+    strictEqual(visitedModels.length, 1);
+  });
 
   it("finds models", async () => {
     const result = await runNavigator(`